requirements.txt
redis>=4.0.0
python-dotenv>=1.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
rank_bm25>=0.2.2
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.1.0
